Rename profile sampling "rate" to "interval"
The profile sampling "rate" denotes the number of bytes between allocations:
this indicates an interval (or inverse rate) but not a rate. The real rate can
be computed as "1/sample_interval" with units "samples/byte".

Rename all variables and functions related to profile sampling "rates" that are
actually intervals in bytes. The old function names in MallocExtension become
aliases to the new names, and are planned to be removed eventually.

No functional change intended.

PiperOrigin-RevId: 642614257
Change-Id: Ibaf0f4811864f384e9180eee8e5a53657031a180