Unconditionally enable subrelease for the cold heap
This makes the hpaa_cold_subrelease flag retired and unconditionally
enables subrelease on the cold heap.

I had to make one of the tests slightly fancier to avoid getting
confused by the presence of the "subrelease enabled" info showing up
in non-subrelease configuration mallocz. Aside from that, this is
pretty straightforward.

PiperOrigin-RevId: 644518940
Change-Id: Ie60c1ab1f3e12d5a33b3218bcd23038b922136d7