Unconditionally enable subrelease for the cold heap
This makes the hpaa_cold_subrelease flag retired and unconditionally
enables subrelease on the cold heap.

I had to make one of the tests slightly fancier to avoid getting
confused by the presence of the "subrelease enabled" info showing up
in non-subrelease configuration mallocz. Aside from that, this is
pretty straightforward.

PiperOrigin-RevId: 644535949
Change-Id: I79a7048f191a930f99fcb48a9fa691e391a1f919