Add a bit more randomness to the sampler, and fix sampling fragment test
This change adds clock_gettime tv.sec + tv_nsec to the sampler seed, improving upon the randomness of the sampler beyond the current 'this' + counter. With the increases randomness it also shows the fragmentation test being more flaky as it checks the fragmentation being within tolerance at the start of the test. The start of the test often has a limited number of samples, creating a larger discrepency between the weighted approximation of the fragmentation and the ones computed from the samples (which also includes local vector and other allocations). This change simple drops the less relevant 'start_profiled' comparison.

PiperOrigin-RevId: 648528067
Change-Id: I5748f666f77c83735fe4232e52e3741c8b50b351