GPA: Fix profile test
The RateLimited test case has been incorrect: it used the same sample interval
for both guarded and normal sampling. This would imply that every sampled
allocation is also a guarded allocation. Fix it by ensuring that the guarded
sampling interval is greater than the profile sampling interval.

This also shows that the current rate limiting logic is subtly broken, because
it would rate limit despite having requested that every sampled allocated is
also guarded. A new test case "NeverRateLimited" demonstrates that. This will
be fixed in a later change.

Furthermore, it used to call GuardedPageAllocator::Reset() for no obvious
reason, besides forcing the test to pass. Remove it.

Instead, we should try to set up each test case to start with similar initial
state. Remove the unused StackTraceFilter::Reset() functions, which are not
suitable when later removal from the filter is possible, and instead introduce
DecayingStackTraceFilter::DecayAll(). Use it in GuardedPageAllocator::Reset(),
and call that before the allocations to be profiled.
PiperOrigin-RevId: 652456775
Change-Id: I5df44345c224b4b9561efb2dd23708da337be50e