Improve sampler random seed
For the level of randomness that the sampler needs, using the CPU's cycle clock
as the initial seed is more than sufficient. Simply switch to using the CPU
cycle clock provided by Abseil, which is already used elsewhere in the TCMalloc
codebase. CycleClock::Now() is defined to be async-signal safe.

As in 92111ef0f079, the sampling test is still flaky. Increase the tolerance of
the starting fragmentation check, which includes allocations that happened before
the test case.

PiperOrigin-RevId: 653251971
Change-Id: Ic5554b80367c44c959777519d4711404eb0c4b7e