Record lifetime of page trackers in huge page filler.
This records lifetime histogram of page trackers in HugeFiller, that prints number of hugepages in filler with lifetime < 1ms, < 10ms, < 100ms, < 1s, < 10s, < 100s, < 1000s and > 1000s. It also prints histogram of number of page allocations on the hugepages that are long-lived (i.e. > 100s lifetime). Additionally, it prints histogram of lifetimes of hugepages with < 12.5% allocated pages (e.g. < 32 allocated pages from 256 pages).

PiperOrigin-RevId: 656006084
Change-Id: Id9fab93adf856bc0085286afd68d89c1169cc6c9