Reduce the number of lists used for dense spans in the HugePageFiller
This is an experimental change.  We know that dense spans only use one TCMalloc
page each.  So hugepages with dense spans do not need to maintain longest free
range and chunks to pick the best possible candidate for a new span allocation.
In this change, we use kPagesPerHugepage length array of linked lists for
holding hugepages with dense spans.  Index 0 is for holding hugepages with all
pages allocated, index 1 for kPagesPerHugepage-1 pages allocated, and so on.

PiperOrigin-RevId: 666047833
Change-Id: I58daa5678b06c0f25338e4e21a825d5c4f5268f5