Performance improvement when computing CacheTopology.
Currently we read the "shared_cpu_list" sysfs file for each CPU in the system, and use it to compute its l3 cache mapping.

Unfortunately these files are protected by a global kernel lock, which can introduce severe contention on large multicore servers when starting several binaries in parallel.

To reduce the number of file reads, we can parse the contents of the shared_cpu_list file into a CpuSet and update l3_cache_index_ for every CPU listed in there.

PiperOrigin-RevId: 670680322
Change-Id: Iea55c706f111d5eb9cce49d7d6ca2289f456b9b3