Reduce cache size in span from 12 to 8 when big span experiment is enabled.
When TEST_ONLY_TCMALLOC_BIG_SPAN experiment is enabled, we increase span size from 48B to 64B. Additionally, we also increase cache size in span from 4 to 8. The additional 8B are reserved. When the experiment is disabled, kCacheSize is equal to 4, as before. We can increase cache size even further in the future to make use of 8B slack as the experiment completes, but this change also allows us to use additional bytes for storing other span metadata.

PiperOrigin-RevId: 671110988
Change-Id: I908ec006ca32b440a113cc60460967e0dab47ef6