Remove a reference to a removed field
The reason why this test has been passing is because the regex is searching for
either `tcmalloc_filler_chunks_per_alloc: 8` or `16`, and the buffer
sufficiently often contains `16` that it continued mostly passing even once the
field was removed. When the output didn't contain `16`, we just considered the
test to be flaky.

PiperOrigin-RevId: 673403378
Change-Id: Ifa00ad2c7248b16dfb057c577b0a05c93e4a750b