memory_errors_test: Better hide deliberate UB from compiler
More recent versions of Clang are able to figure out that in the deliberate
buffer-underflow test an underflow is happening, and turn that into code that
does not crash with an access to GWP-ASan memory. For example by turning the
access into an access to 0xffffffffffffffff, which also crashes but not with
the expected error message.

Fix that by attempting to better hide the underflowing access from the
compiler.

PiperOrigin-RevId: 675173241
Change-Id: Ib43bd4f681799ffcd882ce93d6c4a32166270226