Use fewer bits for num_pages in span.
This reduces the number of bits we use to track num_pages from 64b to 6b, with an additional bit to track if the span allocates larger than kMaxSize object, so a total reduction of 57b. Note this only reduces span size when span is tracked in central freelists. When > kMaxSize object is allocated or if the span is sampled, we track num_pages as a 64b value. The change would effectively help us increase cache and bitmap size in the future.

PiperOrigin-RevId: 677844812
Change-Id: I4e6a1cd6b7aacd1941c8670ece72dff3beac3c69