Move --define='absl=1' and --test_env="GTEST_INSTALL_FAILURE_SIGNAL_HANDLER=1" from CI scripts to .bazelrc
//tcmalloc/internal:logging_test was failing without --define='absl=1'
because the absence of that flag was causing Google Test to link
against the system regular expression library instead of RE2, and
the former has different semantics, at least on my machine ("\\n"
does not match a newline). To avoid these types of problems, let's
make it so that we always link against RE2.

Also AllocationGuard.CooperativeDeathTest was checking for
"SIGABRT received" message printed by the signal handler installed
by GTEST_INSTALL_FAILURE_SIGNAL_HANDLER=1. Since there doesn't
seem to be another reasonable way of checking for SIGABRT from the
GTest side, move that from the CI scripts to .bazelrc as well.

PiperOrigin-RevId: 693756531
Change-Id: I56af3a567e37c72ac0f3597a45b18417fe6830fc