Detect large deallocation sizes inapproriate for their span.
This can indicate data corruption.  For example, on a kMaxSize-sized or smaller
allocation, any of the 45 most significant bits being flipped would push the
deallocation to be treated as a "large" deallocation (do_free_with_size ->
InvokeHooksAndFreePages).  While sampled spans are protected against mismatch,
no such checks are done for unsampled ones.

Sampling is helpful for telling us the allocation stack, but we can detect a
caller-provided size that is larger than the span we are deallocating into.

PiperOrigin-RevId: 698557659
Change-Id: I520423cc091f22b64beefd004eba6e7352608323