Clarify use of static inline functions in headers
In C++ it generally does not make sense to use static inline functions in headers.
If a function is outlined, 'inline' indicates that the function should be
deduplicated at link time (via COMDAT) and we do not end up with multiple
copies.

However, for the per-CPU "static inline" helpers we need "static" to avoid
ODR violations due to different labels emitted for the prologue for each
copy of the function.

PiperOrigin-RevId: 702759097
Change-Id: Icc82faadf30aa937e50a791c35a1de2d63dd77f7