When reporting sized-delete mismatch, refine size bounds.
Many "large" sizes will be clearly erroneous due to bit flips in their most
significant bits, so we do not need a lot of precision to detect the error.
Once we are committed to crashing the process, though, we can lookup if the
span is a size class-ful one and report a narrower range on what permissible
sizes would have been.

While TCMalloc doesn't guarantee the exact text of its error messages, we
inspect it to make sure the size appears in that text to show we did refine the
range.

PiperOrigin-RevId: 702816092
Change-Id: I511b5fa20a221be9acb325096e44528151fa4778