Deflake CpuCacheTest.Metadata.
If we are preempted during the allocation but returned to the same core,
our CPU mask may be unchanged (!Tampered()), but we fail to refill in full and
don't touch the slab.

To avoid this, make multiple attempts and confirm that we were preempted by
verifying the slab is in the uncached state on each failed attempt.

PiperOrigin-RevId: 703845482
Change-Id: I998a1015aa05d5a8ab1d7a468a9edb07e776d02b