Remove unnecessary asserts
The GetSizeClass function was computing the size class in the same way
as do_free (and was only being called from do_free), so the assert
wasn't checking anything useful. Also, we already asserted that ptr
is non-null a few lines above. Therefore, remove both asserts and
the GetSizeClass function.

PiperOrigin-RevId: 706892885
Change-Id: I28ede4a26c1dcf11dd5548cfde942053adfd2b24