Release from HugeCache when limit is hit.
When demand-based memory release is enabled in HugeCache, we might hold on to some free memory. When the limit is hit, we should unconditionally release this memory. This change also adds a test that checks that we indeed release memory from HugeCache when limit is hit.

Fuzz test was unable to catch this as we didn't account for bytes in HugeCache as releasable bytes. This also fixes that.

PiperOrigin-RevId: 720719278
Change-Id: I3269946845d624d35463c7778e6f01d553e1cd59