Fix peak sampled bytes computation.
Peak backed bytes doesn't imply peak sampled bytes. It's possible that the objects are freed and backed in some HPAA component, so, the sampled application bytes would be lower when we have a new peak.

peak_heap_profiling_test is flaky when huge cache demand-based release experiment is enabled. Earlier, we might unback from HugeCache as soon as the memory is freed, so this check never triggered. But now, if we might keep backed memory in HugeCache, so the overall peak upon future allocations might be higher, but total sampled bytes might not peak.

This also deflakes huge_cache_test.

PiperOrigin-RevId: 723724376
Change-Id: I87aaeb87c0a8ebca8823d851998d1d2d50dd38e1