Detect and report potential double frees at the page heap.
When we donated part of our 2MB multiple allocation, we should always
successfully find a page tracker.  When we don't, we almost certainly incurred
a double free.  Direct tests for this condition are omitted, as HPAA's own
tests have a lot of consistency checks that fail first.

In terms of costs, we're already paying for a (predicted) branch for the
TC_CHECK, so this should add no appreciable cost.

PiperOrigin-RevId: 756324935
Change-Id: I52b37ae321d674b108e9216be0e9fc9500b3171a