Userspace hugepage collapse.
This feature builds a periodic userspace hugepage collapse mechanism in TCMalloc's HugePageFiller. Every five seconds, the background thread scans up to 64 page trackers. It choses trackers that:
1. Were not previously recorded to be hugepage backed or successfully collapsed.
2. Were not scanned before (i.e. any newly allocated trackers)
3. Were inspected more than kRecordInterval (5 mins) ago.

For all the chosen trackers, it releases the pageheap lock and then queries pageflag to know if the page is hugepage backed. For pages that aren't hugepage backed, it queries residency information to obtain the number of unbacked and swapped pages in that range. If the number of unbacked and swapped pages are fewer than 64 and 128 respectively, it considers the page eligible for collapse. It then attempts to MADV_COLLAPSE all the eligible pages.

Finally, it reacquires the pageheap lock and updates hugepage, unback and swap state in page trackers for future analysis.

We have to deal with two types of interleaving operations in this scenario considering the pageheap lock is released at multiple instances.
1. Any interleaving subrelease/release operation may either delete the tracker entirely or make the pages partially released. We do not want to collapse such pages. We maintain being_collapsed state in page trackers that is set right before we attempt to collapse and is reset after we collapse. Any interleaving subrelease operation skips the trackers marked being_collapsed.
2. Any interleaving Put operation may deallocate all the pages in the tracker and the tracker may be deleted. We mark all the eligible trackers with a maybe_collapsed state under pageheap lock before it is attempted to be collapsed. Any interleaving Put operations, upon seeing maybe_collapsed state set, and when all the pages within the tracker are deallocated, temporarily store these fully freed trackers in a separate tracker list. Once collapse operations are done, the background thread finally inspects the fully-freed tracker list and deletes all those trackers. We might collapse these pages, only to release them back to the system, in this scenario. But such cases should be rare.

There are a lot of tunable knobs here that we will revisit in the future, but currently, we chose a fairly logical configurations for this mechanism.

PiperOrigin-RevId: 756500290
Change-Id: Ibe60c333a985b4998f8c2b0a11d4bd82132f76f8