Fix coarse LFR indexing.
When LFR bins are coarsely indexed, for an allocation request, the indexing mechanism currently upsizes the index. For instance, for an LFR reqeust of 35, the relevant indexes with the matching trackers either track LFRs [32, 39] or [40, 47]. During allocation, it returns a [40,47] index as it is guaranteed to have trackers with LFRs > 35. However, it does the same thing for a request of 32 LFR. Instead, it can return a [32, 29] bin index.

PiperOrigin-RevId: 758756195
Change-Id: Ib520f53481cab88b5afe087b004db71acbd6d4db