Sample trackers and encode its features in naming their allocated region.
This samples 1% of the allocated trackers when the trackers are newly allocated. Periodically, it iterates through these sampled trackers, records their features (such as the longest free range, number of allocations, etc.), encodes them in a string, and uses that to name the memory tracked by that tracker. Per iteration, it iterates through (up to) 64 trackers. The syscall to prctl() might be expensive, so it releases a page heap lock to make sure we do not call it in critical path.

PiperOrigin-RevId: 776216456
Change-Id: If28282ae8d9368ed301ad5b1e1ba83c7d2b04644