Periodically record features for sampled trackers in HugePageFiller.
This periodically, once every 5 secs, iterates from sampled trackers in
HugePageFiller, records a feature set (such as LFR, number of allocations,
etc.), encodes them in a string and uses that to name the memory tracked by
that tracker. It uses Linux kernel's named vma feature to name the memory
regions.

PiperOrigin-RevId: 776580632
Change-Id: I03f04f029d270db5147d30112778f4f5b84d9a1e