Consolidate hugepage treatments.
This consolidates sampling and collapse treatments applied to page trackers in HugePageFiller. This optimizes the two mechanisms as well - instead of iterating through the tracker lists twice (or perhaps multiple times as we add more treatments in the future), it iterates through the lists once, gathers eligible trackers, and applies treatments (e.g. userspace hugepage collapse) to them.

PiperOrigin-RevId: 778063804
Change-Id: I78e7e5ed29c1aa0284b1b89d529e31f0013121f3