Release free and swapped pages from non-hugepage-backed trackers.
This mechanism, if enabled, periodically inspects up to 64 hugepage trackers in HugePageFiller. For the hugepage-sized regions that aren't hugepage backed:
(1) It collects the number of pages that are unbacked and swapped.
(2) <not part of this change> based on the heuristics (i.e. if number of swapped and unbacked pages are lower than 128 and 64 respectively) used for userspace collapse, it attempts to collapse these hugepages.
(3) In case the hugepage-sized regions aren't eligible for collapse or if the collapse operation fails, if the region has any swapped pages within the range, it attempts to release all the free pages. The idea here is that, given the page isn't hugepage-backed anyway, the performance cost associated with subreleasing from that region is minimal. We might as well reclaim any of the free memory in that region.

PiperOrigin-RevId: 786422484
Change-Id: I091b5ee568600197b20e6df7a157cccc5503d323