Enable userspace hugepage collapse.
This feature builds a periodic userspace hugepage collapse mechanism in TCMalloc's HugePageFiller. Every five seconds, the background thread scans up to 64 page trackers. It choses trackers that:
1. Were not previously recorded to be hugepage backed or successfully collapsed.
2. Were not scanned before (i.e. any newly allocated trackers)
3. Were inspected more than kRecordInterval (5 mins) ago.

For all the chosen trackers, it releases the pageheap lock and then queries pageflag to know if the page is hugepage backed. For pages that aren't hugepage backed, it queries residency information to obtain the number of unbacked and swapped pages in that range. If the number of unbacked and swapped pages are fewer than 64 and 128 respectively, it considers the page eligible for collapse. It then attempts to MADV_COLLAPSE all the eligible pages.

PiperOrigin-RevId: 789088262
Change-Id: I7293924efd0146afc12fddf876e6392e0d80d2ae