Parallel collapse test.
This performs hugepage treatment operations on a parallel thread, while concurrently deallocating pages to previously allocated hugepages. This creates a scenario where we end up with a few trackers in fully freed lists that weren't released (e.g. due to concurrent collapse operations on those freed hugepages). It makes sure that we are handling that scenario right without any crashes or failures.

The test is currently disabled, but would be enabled in the subsequent change with a fix.

PiperOrigin-RevId: 794759872
Change-Id: I39d1b7b327c8e823ca056983ba2b886965ed8c4e