Handle fully freed trackers in RemoveFromFillerList.
While treating trackers in HugePageTreatment, we collect trackers under the lock, but release pageheap lock for gathering certain stats. Another thread might free all pages, and add that tracker to the fully freed list. When we reacquire the lock and attempt to subrelease pages from it, we might want to remove that tracker, and handle this case gracefully in RemoveFromFillerList.

PiperOrigin-RevId: 794772863
Change-Id: I1aae1a828b247af802ebb7bc8746f63a5c1970bb