Ensure deleting destructor does not invoke operator delete.
These are emitted due to the ABI around virtual methods, but are never invoked
in practice by TCMalloc.  Since the vtable is used, it is tricky for the linker
to drop them.

This facilitates analysis to show no TCMalloc function is reentrant for
allocations and provides a dynamic check that we do not use these.

PiperOrigin-RevId: 799689649
Change-Id: Id46e47612ffb2132c4d6a538617669b1de72acc5