Fix metadata_bytes stat computation.
report_residence = false is meant to reduce latency impact due to calculation of actual residency bytes when not required, but it's not meant to report non-resident bytes. When reporting generic.physical_memory_used property, we add non-resident bytes to metadata_bytes twice, once here and then the second time in PhysicalMemoryUsed.
PiperOrigin-RevId: 803162499
Change-Id: I2725ccc0c0aae9558f5111452d65e30d4c33744a