NFC: Minimize stack usage of HugePageFiller telemetry.
This significantly minimizes the amount of state we need to record on stack for printing telemetry. Instead of tracking histograms all at once for all types of trackers, it records histograms individually for each tracker type and prints them so that we do not have to indefinitely keep that state. This change reorders telemetry -> we now print all histograms for one type of tracker at once. The changes in huge_page_filler_test involve reorganizing lines alone, but they do not change the actual information in the telemetry that is being printed out.

PiperOrigin-RevId: 811402376
Change-Id: I6562d5e6d5d318e7f42a88a39b652f9cffb9265b