Add fuzzer for exercising Span scenarios.
It also simulates UB from double frees, allowing us to validate that TCMalloc
attempts to abort the program.  Since we don't want those successful detections
to be flagged as bugs by the fuzzer, we `longjmp` and simply return.

If we do knowingly double free and don't return from `setjmp`, we failed to
detect the bug.

This currently focuses on double frees being detected when the span is fully
drained, but this technique can be extended to check for whether immediate
detection occurs.  Span currently lacks sufficient redundancy for this in the
compressed list representation.

PiperOrigin-RevId: 836857725
Change-Id: Ic5c520be5d9ade8c75e8b0b2b92cea4eafb5454e