Forward partitioning information in non-normal free & realloc.
In a follow up change, partitioning by alloc-token [0]
(pointer-containing/pointerless split) will be supported. In this case, only
pointerless allocations will be put into the cold allocation, hence, we need
to include the partitioning information when determining the size class in the
non-normal free.
For reallocations, we want to keep the same partition too, hence, for cold/hot
sake, we need to add the partition information as well.

[0] https://clang.llvm.org/docs/AllocToken.html

PiperOrigin-RevId: 837850719
Change-Id: Ic0255740fc61066cb79f42e482ea4dc45ffd1119