Separate Span object insertion operations that do not require locks.
Converting from Span's internal ObjIdx representation does not require holding
locks, as it is a function of constant properties of the Span itself or state
uniform across all of the Spans (retained by the CentralFreeList).

Since the Span has objects being returned to it, it cannot become free (and
thus be reused) during this calculation.

PiperOrigin-RevId: 844828662
Change-Id: I8c3db4ca814964135fd195d635cb180a51efb55e