Optimize fast paths for aarch64.
We currently have several instructions used to compute the address of each thread local variable. This cost is reduced here, by changing references to thread local variables after the first one to be relative to the first reference. In the Pop() for example, we need to reference TLS for the sampler, but to access the slabs, we compute an offset from the sampler address, rather than doing a lookup in TLS again.

A few more instructions are saved in the Push/pop rseq inline assembly sequences by using efficient add+shift instructions, pre-indexed ldp, and assuming Top Byte Ignore is enabled.

PiperOrigin-RevId: 846866008
Change-Id: I3134c2d17f196221333be5f3ce1f998e893c100f